Base class and interface
* Added a new base Class and Interface for Deltas, not happy at the moment with the design.
